use `hasuraSdk.JSONValue` type instead of `Record`, `object` and `void` (#39)
This PR replaces `Record`, `object` and `void` types with
`hasuraSdk.JSONValue`. It does that because those types require
`allowrelaxedtypes`, and `void` types caused issues if the API did have
a return value

### Why do we want this?
APIs not defining their parameter types or return types is a common
occurrence in the Open API world. We've run into situations where users
mistake the behaviour of our connector as bug, but the real problem was
that their Open API Document had issues. Also, all of the types
mentioned above require `@allowrelaxedtypes` annotation, and
`hasuraSdk.JSONValue` is a safer way of implementing ambiguous types
that results in less opaque scalars.

The changes are made in
[api-ts-generator](https://github.com/hasura/ndc-open-api-lambda/pull/39/files#diff-5e5e394c0d1b4a09a2a328b0bf47137ee0bf22b51178cc4a83b043496a4acc3b).
All other changes are updates to the golden files for tests