preserve imports from stale files (#90)
This PR adds support for automatically preserving `import` statements in
`functions.ts` and `api.ts` files

## Why
This was required because if a user had a setup in which they were using
functions from files other than `api.ts` and had `@save` annotation on
those functions, re-running the codegen (introspect with overwrite set
to `true`) would preserve the function, but would lose its import
statement. This meant that the user would have to re-write their import
statements again, which a really bad experience.

## Note
`import` statements are always preserved and organized. They don't need
a `@save` annotation.

## Review
Since this is a relatively large PR, it'll be easier to review it per
commit